requirements.txt
openai==1.12.0
langchain==0.1.0
langchain-community==0.0.20
langchain-openai==0.1.25
python-dotenv==1.0.0
streamlit==1.31.0
tiktoken==0.5.2
numpy==1.26.3
pandas==2.2.0
faiss-cpu==1.7.4
sentence-transformers==2.6.1
